Making sure add / remove calls are applied in the order they were called
Previously if `add` was called and then `remove` was called within the
same batch, the batch could be flushed during the `render` of the `add`
which would cause the `remove` to be applied first. This changed makes
sure that any calls to `add` or `remove` do not start running until all
previous calls have completed.