Add support for can.viewInsert
This updates `html()` so it can accept an object with the can.viewInsert symbol as a property. The `html()` method also accepts an options object as the fourth argument; these options will be passed to the can.viewInsert symbol.

These changes are part of https://github.com/canjs/can-stache/issues/502